refactor: 테스트 패키지 구조 변경
- Fake 저장소 객체의 위치를 이동합니다.
