+++ src/test/java/roomescape/business/service/ReservationServiceTest.java
@@ -9,9 +9,9 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.business.fakerepository.FakeReservationRepository;
-import roomescape.business.fakerepository.FakeReservationThemeRepository;
-import roomescape.business.fakerepository.FakeReservationTimeRepository;
+import roomescape.persistence.fakerepository.FakeReservationRepository;
+import roomescape.persistence.fakerepository.FakeReservationThemeRepository;
+import roomescape.persistence.fakerepository.FakeReservationTimeRepository;
 import roomescape.business.service.ReservationService;
 import roomescape.exception.ReservationException;
 import roomescape.persistence.ReservationRepository;
