+++ src/test/java/roomescape/business/service/ReservationThemeServiceTest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.business.Reservation;
 import roomescape.business.ReservationTheme;
-import roomescape.business.fakerepository.FakeReservationRepository;
-import roomescape.business.fakerepository.FakeReservationThemeRepository;
+import roomescape.persistence.fakerepository.FakeReservationRepository;
+import roomescape.persistence.fakerepository.FakeReservationThemeRepository;
 import roomescape.exception.ReservationThemeException;
 import roomescape.persistence.ReservationRepository;
 import roomescape.persistence.ReservationThemeRepository;
