+++ src/test/java/roomescape/business/service/ReservationTimeServiceTest.java
@@ -12,8 +12,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.business.Reservation;
 import roomescape.business.ReservationTime;
-import roomescape.business.fakerepository.FakeReservationRepository;
-import roomescape.business.fakerepository.FakeReservationTimeRepository;
+import roomescape.persistence.fakerepository.FakeReservationRepository;
+import roomescape.persistence.fakerepository.FakeReservationTimeRepository;
 import roomescape.exception.ReservationTimeException;
 import roomescape.persistence.ReservationRepository;
 import roomescape.persistence.ReservationTimeRepository;
